requirements.txt
assertpy==1.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
iniconfig==1.1.1
packaging==21.3
pip==22.0.4
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
requests==2.27.1
setuptools==61.3.1
tomli==2.0.1
urllib3==1.26.9
wheel==0.37.1
playwright==1.21.0
waiting==1.4.1
pytest-playwright==0.3.0
allure-pytest==2.7.1
loguru==0.6.0
